2025-12-6
An achievement feature for game over has been added. The principle is to count the number of vehicles launched by the spawner, and when the number of launched vehicles reaches 0, iterate through the number of vehicles in the scene. If both the number of launched vehicles and the number of launched vehicles are zero, the game ends. Then, a ratio is calculated based on the number of collisions recorded from the vehicle collision script and the total number of launched vehicles, resulting in three different ending UIs with three different ratios. The logic in `simplemanager` has been modified, retaining all functions except for the game over and scoring functions. Incorrect references to `simplemanager` in other scripts have been corrected. Simple integrity testing has been performed. The initial version has been packaged.